requirements.txt
Django>=1.10.5,<1.19.99
django-bootstrap3>=5.4,<5.4.99
django-post-office>=2.0,<2.0.99
python-social-auth>=0.2,<0.2.99
django-userena>=2.0,<2.0.99
python-twitter>=3.1,<3.1.99
Pillow
requests
